project-2/requirements.txt
selenium==4.39.0
webdriver-manager
pytest==9.0.2
allure-pytest==2.15.2